Nodes are now in their own subdirectory.
